Introduce AssertJ to replace Hamcrest.
AssertJ is the de facto successor of Hamcrest. It uses the type system
to provide better IDE support than Hamcrest and tends to help write
assertions that are more expressive.

In this specific project, the changes are minor.

See http://joel-costigliola.github.io/assertj/ for more details.

Also upgraded parent-pom to latest version.

Change-Id: I07efa6cb40c86c37491e6fcae19c23ca9e3fad5a